Persistent player ID on page reloads
https://trello.com/c/R9w5wZlu/92-letterjam-how-to-maintain-a-persistent-player-id

Approach:

- Use express-sessions. This creates a session as a browser cookie that
persists across page loads.
- Use another library to allow accessing express sessions in SocketIO.
- Write to SQLite DB to persist across server restarts
- Use the `session.id` as the playerID.

--

Beyond this change, I did some refactoring. I moved players into the
`ServerGameState`.

Players are represented via a `Map`, but we can change it to a regular
ol' object `{}` if the `Map` is annoying. You *must* use get/set methods
on the map for it to work.  Otherwise you set properties on the map
object, which isn't what you want. :)